Move `root` into `StoredObject`
