+++ models/__init__.py
@@ -1,6 +1,5 @@
-from models.data import timestamp_representer, timestamp_constructor, TIMESTAMP_REPR_STR,\
-    ROOT, DATA_ROOT, json_to_df, get_candles,\
-    combine_data, read_data, write_data, update_candles
+from models.data import json_to_df, get_candles,\
+     combine_data, read_data, write_data, update_candles
 from models.indicator import Indicator, MAX_STRENGTH
 from models.IndicatorContainer import IndicatorContainer
 from models.trades import Trade, SuccessfulTrade, add_to_df, truncate
